Filename management
Updated how a file name that include directories are mananged. The model name, from which It will create the folders and output files, is based on the base name (i.e., removing directory) up to the first period of the base name. So "X\Y\test.name.txt" will name the model "test".